refactor(rules): split attr_duplication into impl/test and align with TS
- Split attr_duplication.rs → attr_duplication/mod.rs + tests.rs
- Rewrite tests to use full lint pipeline (html_arena → lint)
- Align test IDs with TS: invalid_001, invalid_002, invalid_004
- Add v6_attr_duplication_001..005 for Rust-only tests
- Fix message to match TS: "The attribute name is duplicated"
- Fix raw to use name token raw (preserves original case)
- Assert line:col with exact TS template literal input bytes
